+++ contracts/src/interfaces/ITransfersListener.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity ^0.8.21;
+pragma solidity ^0.8.18;
 
 interface ITransfersListener {
     function onLTokenTransfers(address from, address to, uint256 amount) external;