+++ hardhat.config.ts
@@ -5,7 +5,7 @@ import { HardhatUserConfig } from "hardhat/config";
 
 const config: HardhatUserConfig = {
   solidity: {
-    version: "0.8.21",
+    version: "0.8.18",
     settings: {
       evmVersion: "london",
       optimizer: {